+++ apps/vmq_server/test/vmq_schema_SUITE.erl
@@ -38,7 +38,9 @@ all() ->
     [proxy_protocol_inheritance_test,
      proxy_protocol_override_test,
      ssl_certs_opts_inheritance_test,
-     ssl_certs_opts_override_test].
+     ssl_certs_opts_override_test,
+     allowed_protocol_versions_inheritance_test,
+     allowed_protocol_versions_override_test].
 
 
 ssl_certs_opts_inheritance_test(_Config) ->
@@ -184,17 +186,73 @@ proxy_protocol_override_test(_Config) ->
     true = expect(Conf, [vmq_server, listeners, http,  {{127,0,0,1}, 8888},proxy_protocol]),
     true = expect(Conf, [vmq_server, listeners, mqttws,{{127,0,0,1}, 800}, proxy_protocol]).
 
+allowed_protocol_versions_inheritance_test(_Config) ->
+    Conf = [%% required settings the listener translator.
+            {["listener","max_connections"], "10000"},
+            {["listener","nr_of_acceptors"], "100"},
+            %% tcp/mqtt
+            {["listener","tcp","allowed_protocol_versions"], "[3,4,5]"},
+            {["listener","tcp","default"],"127.0.0.1:1884"},
+            %% tcp/ssl/mqtt
+            {["listener","ssl","allowed_protocol_versions"], "[3,4,5]"},
+            {["listener","ssl","default"],"127.0.0.1:8884"},
+            %% websocket
+            {["listener","ws","allowed_protocol_versions"], "[3,4,5]"},
+            {["listener","ws","default"],"127.0.0.1:800"},
+            %% websocket/ssl
+            {["listener","wss","allowed_protocol_versions"], "[3,4,5]"},
+            {["listener","wss","default"],"127.0.0.1:900"}],
+    [3,4,5] = expect(Conf, [vmq_server, listeners, mqtt,  {{127,0,0,1}, 1884},allowed_protocol_versions]),
+    [3,4,5] = expect(Conf, [vmq_server, listeners, mqtts,  {{127,0,0,1}, 8884},allowed_protocol_versions]),
+    [3,4,5] = expect(Conf, [vmq_server, listeners, mqttws,{{127,0,0,1}, 800}, allowed_protocol_versions]),
+    [3,4,5] = expect(Conf, [vmq_server, listeners, mqttwss,{{127,0,0,1}, 900}, allowed_protocol_versions]).
+
+allowed_protocol_versions_override_test(_Config) ->
+    Conf = [%% required settings on listener translator.
+            {["listener","max_connections"], "10000"},
+            {["listener","nr_of_acceptors"], "100"},
+            %% tcp/mqtt
+            {["listener","tcp","allowed_protocol_versions"], "[3,4]"},
+            {["listener","tcp","default"],"127.0.0.1:1884"},
+            {["listener","tcp","default","allowed_protocol_versions"], "[4]"},
+            %% tcp/ssl/mqtt
+            {["listener","ssl","allowed_protocol_versions"], "[3,4]"},
+            {["listener","ssl","default"],"127.0.0.1:8884"},
+            {["listener","ssl","default","allowed_protocol_versions"], "[4]"},
+            %% websocket
+            {["listener","ws","allowed_protocol_versions"], "[3,4]"},
+            {["listener","ws","default"],"127.0.0.1:800"},
+            {["listener","ws","default","allowed_protocol_versions"], "[4]"},
+            %% websocket/ssl
+            {["listener","wss","allowed_protocol_versions"], "[3,4]"},
+            {["listener","wss","default"],"127.0.0.1:900"},
+            {["listener","wss","default","allowed_protocol_versions"], "[4]"}
+           ],
+    [4] = expect(Conf, [vmq_server, listeners, mqtt, {{127,0,0,1}, 1884}, allowed_protocol_versions]),
+    [4] = expect(Conf, [vmq_server, listeners, mqttws, {{127,0,0,1}, 800}, allowed_protocol_versions]),
+    [4] = expect(Conf, [vmq_server, listeners, mqttws,{{127,0,0,1}, 800}, allowed_protocol_versions]),
+    [4] = expect(Conf, [vmq_server, listeners, mqttwss,{{127,0,0,1}, 900}, allowed_protocol_versions]).
+
+
+-define(stacktrace, try throw(foo) catch foo -> erlang:get_stacktrace() end).
+
 expect(Conf, Setting) ->
     Schema = cuttlefish_schema:files([code:priv_dir(vmq_server) ++ "/vmq_server.schema"]),
-    Gen = cuttlefish_generator:map(Schema,Conf),
-    deep_find(Gen, Setting).
+    case cuttlefish_generator:map(Schema,Conf) of
+        {error, _, _} = E ->
+            StackTrace = ?stacktrace,
+            throw({E, StackTrace});
+        Gen ->
+            deep_find(Gen, Setting)
+    end.
 
 deep_find(Value, []) ->
     Value;
 deep_find(Conf, [Prop|T]) ->
     case lists:keyfind(Prop, 1, Conf) of
         false ->
-            throw({could_not_find, Prop, in, Conf});
+            StackTrace = ?stacktrace,
+            throw({could_not_find, Prop, in, Conf, StackTrace});
         {Prop, Value} ->
             deep_find(Value, T)
     end.